021.slack-lambda-mcp-server/scripts/py/requirements.txt
boto3==1.28.38
requests==2.32.4
python-dotenv==1.0.0